improve custom style
