+++ src/full-text-present/CustomStyle.scss
@@ -1,5 +1,9 @@
 @import '../others/variables.scss';
 
 .custom-style {
-  
+    background-color: var(--bs-gray-800) !important;
+
+    .custom-style-body {
+        background-color: var(--bs-gray-700) !important;
+    }
 }
