+++ src/slide-presenting/Presenting.tsx
@@ -2,7 +2,7 @@ import './Presenting.scss';
 
 import SlidePreview from './SlidePresentingController';
 import FullTextPresentController from '../full-text-present/FullTextPresentController';
-import { useBiblePresenting } from '../event/FullTextPresentEventListener';
+import { useFullTextPresenting } from '../event/FullTextPresentEventListener';
 import { useSlideItemThumbSelecting, useSlideSelecting } from '../event/SlideListEventListener';
 import fullTextPresentHelper from '../full-text-present/fullTextPresentHelper';
 import { useStateSettingString } from '../helper/settingHelper';
@@ -10,7 +10,7 @@ import { useStateSettingString } from '../helper/settingHelper';
 export default function Presenting() {
     // s: slides, f: full text
     const [tabType, setTabType] = useStateSettingString('slide-presenting-tab', 's');
-    useBiblePresenting(() => setTabType('f'));
+    useFullTextPresenting(() => setTabType('f'));
     useSlideSelecting(() => setTabType('s'));
     useSlideItemThumbSelecting(() => setTabType('s'));
     return (
